refactor(wolfdesk): consolidate schemas into single file with shared field groups
Merge ticket.action.schemas.ts, ticket.event.schemas.ts, and
ticket.state.schemas.ts into ticket.schemas.ts. Extract shared field
groups (OpenTicketFields, MessageFields, AssignmentFields,
EscalationFields) to eliminate duplication between action/event pairs
and use .extend() for partial state composition.
