+++ packages/wolfdesk/src/schemas/index.ts
@@ -1,4 +1,2 @@
 export * as external from "./external.schemas.js";
-export * as actionSchemas from "./ticket.action.schemas.js";
-export * as eventSchemas from "./ticket.event.schemas.js";
-export * from "./ticket.state.schemas.js";
+export * from "./ticket.schemas.js";
